requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0.0,>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
tqdm>=4.64.0
tensorboard>=2.10.0
scikit-image>=0.19.0
opencv-python>=4.6.0
flask>=2.3.0
werkzeug>=2.3.0
requests>=2.28.0

